Refactor dicts containing lists or single elements
